+++ NEWS.md
@@ -6,6 +6,7 @@
 
 Bugs:
 
+* `plot_predictions(points=1)` errors with `y|trials(n)~` formulasThanks to @avehtari for report #1615.
 * Custom functions in `hypothesis` formulas now correctly capture their parent environment, allowing them to access user-defined variables and weights. Thanks to @strengejacke for report #1608.
 * Fix to `get_coef.betareg()` to extract `precision` parameter instead of just `phi`. This fixes a mismatch issue with names in the vcov. Thanks to @ngreifer for the report.